Slight fix to covariance test to ensure doubles are used.

+++ src/cormorant/tests/test_cg_product.py
@@ -99,7 +99,7 @@ class TestCGProduct():
 
 def _gen_rot(angles, maxl):
     alpha, beta, gamma = angles
-    D = rot.WignerD_list(maxl, alpha, beta, gamma)
+    D = rot.WignerD_list(maxl, alpha, beta, gamma, dtype=torch.double)
     R = rot.EulerRot(alpha, beta, gamma)
 
     return D, R
